:sparkles: Added functional and class approach as example
In order to keep clean the entry point of our application, now starter examples show the way to create classes and use it in our index.ts file. Also was improved the way in were functional approach use functions in order to keep clean undestanding of which functions should be executed and which ones not.